Modulized endpoints into their respective nested routes
